rename util files
