+++ client/src/Components/Roster/GuildMemberCard.tsx
@@ -14,8 +14,8 @@ import DiscordRole from '../../interfaces/discord-role';
 import MemberRecord from '../../interfaces/member-record';
 import { useUpdateDiscordMember } from '../../utils/apis/discord-api';
 import { useAddWarningMutation } from '../../utils/apis/warnings-api';
-import { getDateString } from '../../utils/DataProcessing';
-import { getColorFromRole } from '../../utils/Helpers';
+import { getDateString } from '../../utils/data-processing';
+import { getColorFromRole } from '../../utils/helpers';
 import './GuildMemberCard.scss';
 import GuildMemberDetails from './GuildMemberDetails';
 import GuildMemberMenu from './GuildMemberMenu';
