+++ client/src/utils/helpers.ts
